+++ esp-hal/src/uart.rs
@@ -1222,6 +1222,12 @@ where
 
         self.rx.uart.info().apply_config(&config)?;
 
+        // Don't wait after transmissions by default,
+        // so that bytes written to TX FIFO are always immediately transmitted.
+        self.register_block()
+            .idle_conf()
+            .modify(|_, w| w.tx_idle_num().bits(0));
+
         // Setting err_wr_mask stops uart from storing data when data is wrong according
         // to reference manual
         self.register_block()